remove FormTemplate
