+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\dater-test\\dater-test-game\\src\\index.tsx":"1","D:\\dater-test\\dater-test-game\\src\\App.tsx":"2","D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\index.tsx":"3","D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\NameForm.tsx":"4","D:\\dater-test\\dater-test-game\\src\\store\\reducer.ts":"5","D:\\dater-test\\dater-test-game\\src\\store\\types.ts":"6","D:\\dater-test\\dater-test-game\\src\\store\\actions.ts":"7","D:\\dater-test\\dater-test-game\\src\\pages\\CreateScreen\\index.tsx":"8","D:\\dater-test\\dater-test-game\\src\\components\\FormTemplate\\index.tsx":"9","D:\\dater-test\\dater-test-game\\src\\pages\\CreateScreen\\CreateGameForm.tsx":"10"},{"size":671,"mtime":1608105208082,"results":"11","hashOfConfig":"12"},{"size":593,"mtime":1608155636377,"results":"13","hashOfConfig":"12"},{"size":547,"mtime":1608155718669,"results":"14","hashOfConfig":"12"},{"size":843,"mtime":1608154391890,"results":"15","hashOfConfig":"12"},{"size":598,"mtime":1608156713110,"results":"16","hashOfConfig":"12"},{"size":432,"mtime":1608156721809,"results":"17","hashOfConfig":"12"},{"size":297,"mtime":1608156639579,"results":"18","hashOfConfig":"12"},{"size":275,"mtime":1608156353435,"results":"19","hashOfConfig":"12"},{"size":1473,"mtime":1608157844096,"results":"20","hashOfConfig":"12"},{"size":540,"mtime":1608157014660,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},"gvxkxr",{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\dater-test\\dater-test-game\\src\\index.tsx",[],["43","44"],"D:\\dater-test\\dater-test-game\\src\\App.tsx",[],"D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\index.tsx",[],"D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\NameForm.tsx",[],"D:\\dater-test\\dater-test-game\\src\\store\\reducer.ts",[],"D:\\dater-test\\dater-test-game\\src\\store\\types.ts",[],"D:\\dater-test\\dater-test-game\\src\\store\\actions.ts",[],"D:\\dater-test\\dater-test-game\\src\\pages\\CreateScreen\\index.tsx",[],"D:\\dater-test\\dater-test-game\\src\\components\\FormTemplate\\index.tsx",[],"D:\\dater-test\\dater-test-game\\src\\pages\\CreateScreen\\CreateGameForm.tsx",[],{"ruleId":"45","replacedBy":"46"},{"ruleId":"47","replacedBy":"48"},"no-native-reassign",["49"],"no-negated-in-lhs",["50"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\dater-test\\dater-test-game\\src\\index.tsx":"1","D:\\dater-test\\dater-test-game\\src\\App.tsx":"2","D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\index.tsx":"3","D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\NameForm.tsx":"4","D:\\dater-test\\dater-test-game\\src\\store\\reducer.ts":"5","D:\\dater-test\\dater-test-game\\src\\store\\types.ts":"6","D:\\dater-test\\dater-test-game\\src\\store\\actions.ts":"7","D:\\dater-test\\dater-test-game\\src\\pages\\CreateScreen\\index.tsx":"8","D:\\dater-test\\dater-test-game\\src\\pages\\CreateScreen\\CreateGameForm.tsx":"9"},{"size":671,"mtime":1608105208082,"results":"10","hashOfConfig":"11"},{"size":593,"mtime":1608155636377,"results":"12","hashOfConfig":"11"},{"size":547,"mtime":1608155718669,"results":"13","hashOfConfig":"11"},{"size":1350,"mtime":1608191396669,"results":"14","hashOfConfig":"11"},{"size":598,"mtime":1608156713110,"results":"15","hashOfConfig":"11"},{"size":432,"mtime":1608156721809,"results":"16","hashOfConfig":"11"},{"size":297,"mtime":1608156639579,"results":"17","hashOfConfig":"11"},{"size":275,"mtime":1608156353435,"results":"18","hashOfConfig":"11"},{"size":1180,"mtime":1608191447046,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"gvxkxr",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\dater-test\\dater-test-game\\src\\index.tsx",[],["39","40"],"D:\\dater-test\\dater-test-game\\src\\App.tsx",[],"D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\index.tsx",[],"D:\\dater-test\\dater-test-game\\src\\pages\\StartScreen\\NameForm.tsx",[],"D:\\dater-test\\dater-test-game\\src\\store\\reducer.ts",[],"D:\\dater-test\\dater-test-game\\src\\store\\types.ts",[],"D:\\dater-test\\dater-test-game\\src\\store\\actions.ts",[],"D:\\dater-test\\dater-test-game\\src\\pages\\CreateScreen\\index.tsx",[],"D:\\dater-test\\dater-test-game\\src\\pages\\CreateScreen\\CreateGameForm.tsx",[],{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},"no-native-reassign",["45"],"no-negated-in-lhs",["46"],"no-global-assign","no-unsafe-negation"]
