+++ src/pages/CreateScreen/CreateGameForm.tsx
@@ -1,22 +1,40 @@
-import React from "react"
-import FormTemplate from "../../components/FormTemplate"
+import {Block, List, ListInput, Button} from "framework7-react"
+import React, {useRef} from "react"
+import {useDispatch} from "react-redux"
 import {createGame} from "../../store/actions"
 
-const formCustomStyle = {
+const blockCustomStyle = {
     marginTop: "40vh",
     maxWidth: "40%",
     marginLeft: "30%",
 }
 
 const CreateGameForm: React.FC = () => {
+    const dispatch = useDispatch()
+    const nameRef = useRef<any>(null)
+    const handleConfirm = () => {
+        const name = nameRef.current.__reactRefs.inputEl.value
+        try {
+            if (!name) {
+                throw new Error("Empty string")
+            }
+            dispatch(createGame(name))
+        } catch {}
+    }
     return (
-        <FormTemplate
-            save={createGame}
-            label="Name your game"
-            placeholder="Name"
-            buttonName="Create"
-            style={formCustomStyle}
-        />
+        <Block style={blockCustomStyle}>
+            <List inlineLabels>
+                <ListInput
+                    ref={nameRef}
+                    type="text"
+                    label="Name your game"
+                    placeholder="Name"
+                    autofocus
+                    clearButton
+                />
+            </List>
+            <Button onClick={handleConfirm}>Create</Button>
+        </Block>
     )
 }
